// src/hooks/use-algorand/types.gen.ts
export type ApplicationLocalState = {
	/**
	 * The application which this local state is for.
	 */
	id: number;
	"key-value"?: TealKeyValueStore;
	schema: ApplicationStateSchema;
};

/**
 * Stores the global information associated with an application.
 */
export type ApplicationParams = {
	/**
	 * \[approv\] approval program.
	 */
	"approval-program": string;
	/**
	 * \[clearp\] approval program.
	 */
	"clear-state-program": string;
	/**
	 * The address that created this application. This is the address where the parameters and global state for this application can be found.
	 */
	creator: string;
	/**
	 * \[epp\] the amount of extra program pages available to this app.
	 */
	"extra-program-pages"?: number;
	"global-state"?: TealKeyValueStore;
	"global-state-schema"?: ApplicationStateSchema;
	"local-state-schema"?: ApplicationStateSchema;
};
// ...
export type ApplicationStateSchema = {
	/**
	 * \[nbs\] num of byte slices.
	 */
	"num-byte-slice": number;
	/**
	 * \[nui\] num of uints.
	 */
	"num-uint": number;
};
// ...
export type TealKeyValueStore = Array<TealKeyValue>;
// ...
export type AccountInformationData = {
	/**
	 * An account public key
	 */
	address: string;
	/**
	 * When set to `all` will exclude asset holdings, application local state, created asset parameters, any created application parameters. Defaults to `none`.
	 */
	exclude?: "all" | "none";
	/**
	 * Configures whether the response object is JSON or MessagePack encoded. If not provided, defaults to JSON.
	 */
	format?: "json" | "msgpack";
};

export type AccountInformationResponse = Account;

export type AccountApplicationInformationData = {
	/**
	 * An account public key
	 */
	address: string;
	/**
	 * An application identifier
	 */
	applicationId: number;
	/**
	 * Configures whether the response object is JSON or MessagePack encoded. If not provided, defaults to JSON.
	 */
	format?: "json" | "msgpack";
};

export type AccountApplicationInformationResponse = {
	"app-local-state"?: ApplicationLocalState;
	"created-app"?: ApplicationParams;
	/**
	 * The round for which this information is relevant.
	 */
	round: number;
};
// ...
export type AccountAssetInformationData = {
	/**
	 * An account public key
	 */
	address: string;
	/**
	 * An asset identifier
	 */
	assetId: number;
	/**
	 * Configures whether the response object is JSON or MessagePack encoded. If not provided, defaults to JSON.
	 */
	format?: "json" | "msgpack";
};

export type AccountAssetInformationResponse = {
	"asset-holding"?: AssetHolding;
	"created-asset"?: AssetParams;
	/**
	 * The round for which this information is relevant.
	 */
	round: number;
};

export type GetPendingTransactionsByAddressData = {
	/**
	 * An account public key
	 */
	address: string;
	/**
	 * Configures whether the response object is JSON or MessagePack encoded. If not provided, defaults to JSON.
	 */
	format?: "json" | "msgpack";
	/**
	 * Truncated number of transactions to display. If max=0, returns all pending txns.
	 */
	max?: number;
};

export type GetPendingTransactionsByAddressResponse = {
	/**
	 * An array of signed transaction objects.
	 */
	"top-transactions": Array<unknown>;
	/**
	 * Total number of transactions in the pool.
	 */
	"total-transactions": number;
};

export type GetApplicationByIdData = {
	/**
	 * An application identifier
	 */
	applicationId: number;
};

export type GetApplicationByIdResponse = Application;

export type GetApplicationBoxByNameData = {
	/**
	 * An application identifier
	 */
	applicationId: number;
	/**
	 * A box name, in the goal app call arg form 'encoding:value'. For ints, use the form 'int:1234'. For raw bytes, use the form 'b64:A=='. For printable strings, use the form 'str:hello'. For addresses, use the form 'addr:XYZ...'.
	 */
	name: string;
};

export type GetApplicationBoxByNameResponse = Box;

export type GetApplicationBoxesData = {
	/**
	 * An application identifier
	 */
	applicationId: number;
	/**
	 * Max number of box names to return. If max is not set, or max == 0, returns all box-names.
	 */
	max?: number;
};

export type GetApplicationBoxesResponse = {
	boxes: Array<BoxDescriptor>;
};

export type GetAssetByIdData = {
	/**
	 * An asset identifier
	 */
	assetId: number;
};

export type GetAssetByIdResponse = Asset;

export type GetBlockData = {
	/**
	 * Configures whether the response object is JSON or MessagePack encoded. If not provided, defaults to JSON.
	 */
	format?: "json" | "msgpack";
	/**
	 * The round from which to fetch block information.
	 */
	round: number;
};

export type GetBlockResponse = {
	/**
	 * Block header data.
	 */
	block: unknown;
	/**
	 * Optional certificate object. This is only included when the format is set to message pack.
	 */
	cert?: unknown;
};

export type GetBlockHashData = {
	/**
	 * The round from which to fetch block hash information.
	 */
	round: number;
};
// ...
export type GetTransactionProofData = {
	/**
	 * Configures whether the response object is JSON or MessagePack encoded. If not provided, defaults to JSON.
	 */
	format?: "json" | "msgpack";
	/**
	 * The type of hash function used to create the proof, must be one of:
	 * * sha512_256
	 * * sha256
	 */
	hashtype?: "sha512_256" | "sha256";
	/**
	 * The round in which the transaction appears.
	 */
	round: number;
	/**
	 * The transaction ID for which to generate a proof.
	 */
	txid: string;
};
// ...
export type GetLedgerStateDeltaForTransactionGroupData = {
	/**
	 * Configures whether the response object is JSON or MessagePack encoded. If not provided, defaults to JSON.
	 */
	format?: "json" | "msgpack";
	/**
	 * A transaction ID, or transaction group ID
	 */
	id: string;
};

export type GetLedgerStateDeltaForTransactionGroupResponse = LedgerStateDelta;

export type GetLedgerStateDeltaData = {
	/**
	 * Configures whether the response object is JSON or MessagePack encoded. If not provided, defaults to JSON.
	 */
	format?: "json" | "msgpack";
	/**
	 * The round for which the deltas are desired.
	 */
	round: number;
};

export type GetLedgerStateDeltaResponse = LedgerStateDelta;

export type GetTransactionGroupLedgerStateDeltasForRoundData = {
	/**
	 * Configures whether the response object is JSON or MessagePack encoded. If not provided, defaults to JSON.
	 */
	format?: "json" | "msgpack";
	/**
	 * The round for which the deltas are desired.
	 */
	round: number;
};
// ...
export type WaitForBlockData = {
	/**
	 * The round to wait until returning status
	 */
	round: number;
};
// ...
export type PendingTransactionInformationData = {
	/**
	 * Configures whether the response object is JSON or MessagePack encoded. If not provided, defaults to JSON.
	 */
	format?: "json" | "msgpack";
	/**
	 * A transaction ID
	 */
	txid: string;
};